requirements.txt
alembic==1.13.2
aiogram==3.13.1
SQLAlchemy==2.0.35
python-dotenv==1.0.1
